splitting up code, work in progress
